+++ app/controllers/static_controller.rb
@@ -40,11 +40,6 @@ class StaticController < ApplicationController
   end
 
   def sitemap
-    respond_to do |format|
-      format.html
-      format.xml
-      format.text
-    end
   end
 
   def governance_policy